Created Options package & moved Add. Made things public so that Add can access the other package.
