+++ src/main/java/com/pelensky/contactmanager/Commands.java
@@ -13,7 +13,7 @@ class Commands {
 
     int selectContactTo(String text) {
         io.displayText(
-                "Which contact would you like to" + text + "?"
+                "Which contact would you like to " + text + "?"
                         + System.lineSeparator()
                         + "Please select number.");
         io.displayText(manipulateContacts.listContacts(contactList.getContacts()));
