+++ src/main/java/com/pelensky/contactmanager/IO.java
@@ -13,11 +13,11 @@ public class IO {
     this.output = output;
   }
 
-  String getUserInput() {
+  public String getUserInput() {
     return input.nextLine().trim();
   }
 
-  void displayText(String text) {
+  public void displayText(String text) {
     output.println(text);
   }
